+++ CHANGELOG.md
@@ -2,6 +2,31 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [3.10.0](https://github.com/lifinance/sdk/compare/v3.9.3...v3.10.0) (2025-08-19)
+
+
+### Features
+
+* add getWalletBalances() ([#288](https://github.com/lifinance/sdk/issues/288)) ([1ec6446](https://github.com/lifinance/sdk/commit/1ec6446befa9a9b3f241ba3d535770f8c8c00cc3))
+
+### [3.9.3](https://github.com/lifinance/sdk/compare/v3.9.2...v3.9.3) (2025-08-18)
+
+### [3.9.2](https://github.com/lifinance/sdk/compare/v3.9.1...v3.9.2) (2025-08-13)
+
+
+### Bug Fixes
+
+* update batching conditions to include relayer transactions ([a458d29](https://github.com/lifinance/sdk/commit/a458d29d25959a73c5a4959d23ce5c5f9e6ecc79))
+
+### [3.9.1](https://github.com/lifinance/sdk/compare/v3.9.0...v3.9.1) (2025-08-11)
+
+
+### Bug Fixes
+
+* adjust tx links handling ([#294](https://github.com/lifinance/sdk/issues/294)) ([e4defe1](https://github.com/lifinance/sdk/commit/e4defe152a8a55aab884b010c8d4a0a5b88e590e))
+
+## [3.9.0](https://github.com/lifinance/sdk/compare/v3.9.0-beta.0...v3.9.0) (2025-08-08)
+
 ### [3.8.11](https://github.com/lifinance/sdk/compare/v3.8.10...v3.8.11) (2025-08-05)